+++ src/CmdOptions.cs
@@ -7,4 +7,7 @@ public class CmdOptions
 {
     [Value(0, MetaName = "InputFile", HelpText = "File to parse")]
     public String? InputFile { get; set; }
+
+    [Option('r', "replay", HelpText = "Replay data on serial port", Default = false)]
+    public bool replayOnSerial { get; set; }
 }