+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file holds the Changelog of the project and all relevant changes.
 - Implemented a console printer for formatted output per telegram
 - Implemented TelegramPlayer for replaying telegrams from files
 - First base information for GSM status
+- Implemented SerialSimulator
 
 ## 1.0.0
 Initial release including support for live parsing and parsing of stored files.